// frontend/src/components/game/PingPong/theme.d.ts
export const theme1 = {
	canvas: {
		color: "BLACK"
	},
	player1: {
		color: "#EEEEEE"
	},
	player2: {
		color: "#EEEEEE"
	},
	ball: {
		color: "#EEEEEE",
		radius: 15
	},
	net: {
		color: "#EEEEEE"
	}
}

export const theme2 = {
	canvas: {
		firstColor: "#295c7d",
		secondColor: "#112B3C"
	},
	player1: {
		firstColor: "#F66B0E",
		secondColor: "#f5883f"
	},
	player2: {
		firstColor: "#f5883f",
		secondColor: "#F66B0E"
	},
	ball: {
		firstColor: "#f69b5e",
		secondColor: "#F66B0E",
		radius: 15
	},
	net: {
		firstColor: "#4c81a5",
		secondColor: "#295c7d"
	}
}